+++ backend/__tests__/routes/status.test.js
@@ -6,7 +6,7 @@ const Status = require("../../models/status");
 const testStatus = {
     accountId: 678,
     temperature: 40,
-    statusTime: "2000-09-22",
+    statusTime: "2000-09-22 00:00:00.000",
     isReviewed: "0",
     weight: 100,
     symptoms: "My head hurts",
