--- frontend/components/homepage/types/card.ts
@@ -1,15 +1,0 @@
-export interface CardProps {
-    label: string;
-    image?: string;
-    url:string;
-    style?: Object;
-  }
-
-  export interface CardPatientProps {
-    label: string;
-    temperature: number;
-    weight: number;
-    symptoms: string;
-  }
-
-
