+++ frontend/pages/index.tsx
@@ -5,6 +5,7 @@ import { Box, Img, Flex, Heading } from "@chakra-ui/react";
 import Dashboard from "@frontend/components/homepage/dashboard";
 import { USER_ROLES } from "@frontend/utils/constants";
 import { serverURL } from "@frontend/config";
+import { StatusDataProps } from "@frontend/components/homepage/types/types";
 
 export const getServerSideProps: GetServerSideProps = async context => {
     let session = await getSession(context);
@@ -25,9 +26,11 @@ export const getServerSideProps: GetServerSideProps = async context => {
     };
 };
 
-export default function Home({ data }: any) {
-    const { data: session } = useSession();
+export default function Home({ data }: {data: StatusDataProps}) {
+    console.log( data);
 
+    const { data: session } = useSession();
+    console.log( data);
     if (session) {
         return <Dashboard data={data} />;
     }
